+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- "Latest" navigation button for recent episodes
+
+### Changed
+
+- Episode tile UI improvements
+
+### Fixed
+
+- Socket reconnection on connection status change
+
 ## [v0.5.0] - 2026-07-04
 
 ### Changed